Moved DI for viewmodels
